Document a couple more networking-related classes
