+++ Source/Client/Comp/MapAsyncTime.cs
@@ -5,6 +5,7 @@ using System.Linq;
 using Harmony;
 using Multiplayer.Client.Sync;
 using Multiplayer.Common;
+using Multiplayer.Server;
 using RimWorld;
 using RimWorld.Planet;
 using UnityEngine;
