+++ Source/Client/Multiplayer.cs
@@ -28,6 +28,8 @@ using Multiplayer.Client.Comp;
 using Multiplayer.Client.Networking;
 using Multiplayer.Client.Sync;
 using Multiplayer.Client.Windows;
+using Multiplayer.Common.Networking;
+using Multiplayer.Server;
 using UnityEngine;
 using Verse;
 using Verse.Profile;
