+++ Source/Client/MultiplayerMod.cs
@@ -14,6 +14,7 @@ using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 using System.Text;
 using System.Xml;
+using Multiplayer.Server;
 using UnityEngine;
 using Verse;
