+++ Source/Client/Replays.cs
@@ -7,6 +7,7 @@ using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Text;
+using Multiplayer.Common.Networking;
 using UnityEngine;
 using Verse;
 using zip::Ionic.Zip;
