+++ Source/Client/SteamIntegration.cs
@@ -6,6 +6,7 @@ using System.Diagnostics;
 using System.Linq;
 using System.Text;
 using Multiplayer.Common.Networking;
+using Multiplayer.Server;
 using Multiplayer.Server.Networking;
 using UnityEngine;
 using Verse;
