+++ Source/Client/Windows/HostWindow.cs
@@ -1,6 +1,7 @@
 ﻿using System.Net;
 using Multiplayer.Client.Networking;
 using Multiplayer.Common;
+using Multiplayer.Server;
 using RimWorld;
 using UnityEngine;
 using Verse;
