+++ Source/Common/ChatCommands.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
+using Multiplayer.Server;
 
 namespace Multiplayer.Common
 {
