+++ Source/Server/MultiplayerServer.cs
@@ -1,5 +1,4 @@
-﻿using LiteNetLib;
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.Linq;
@@ -7,9 +6,12 @@ using System.Net;
 using System.Net.Sockets;
 using System.Text;
 using System.Threading;
+using LiteNetLib;
+using Multiplayer.Common;
+using Multiplayer.Common.Networking;
 using Verse;
 
-namespace Multiplayer.Common
+namespace Multiplayer.Server
 {
     public class MultiplayerServer
     {
